Tabular Data Classification (Ranked 5th on Kaggle)
Tabular Data. Multi-Class Classification. The dataset is the Airbnb listings of different areas in Montreal during 2019. It comes with rich information for each listing, including a link to the thumbnails etc.  We build models to predict the listing price based on the listing characteristics. In this way, we can optimize the user experience and lower the bar to be a new host😍 ! 

Predicting the actual price could be simple (you can just find it online), so we cut the pricing into three different bins for classification 😂. For each listing, we recommend a pricing range to the new host rather than a fixed price (how nice is that!). So we define three categories: beginner, plus, premium based on the created listing. Respectively we use 0, 1, 2 to denote these three categories.

The data can be viewed&retrieved from https://www.kaggle.com/c/cisc372-2021w-a2/data